fix conversion warnings on posix

+++ include/boost/mmap/flags/posix/mapping.hpp
@@ -57,8 +57,8 @@ struct viewing
     bool operator< ( viewing const other ) const noexcept
     {
         return
-            ( ( other.protection & access_privileges::write   ) && !( this->protection & access_privileges::write   ) ) ||
-            ( ( other.protection & access_privileges::execute ) && !( this->protection & access_privileges::execute ) );
+            ( ( static_cast< std::uint32_t >( other.protection ) & access_privileges::write   ) && !( static_cast< std::uint32_t >( this->protection ) & access_privileges::write   ) ) ||
+            ( ( static_cast< std::uint32_t >( other.protection ) & access_privileges::execute ) && !( static_cast< std::uint32_t >( this->protection ) & access_privileges::execute ) );
     }
 
     bool operator<=( viewing const other ) const noexcept
